+++ tests/test_reconstructions.py
@@ -1,8 +1,8 @@
 """
 Test the reconstruction of the IFU head
 """
-from __future__ import (absolute_import, division, print_function,
-                        unicode_literals)
+from __future__ import (absolute_import, division, print_function)
+# , unicode_literals)
 
 import pytest